Improve the rendering a little bit
The idea was to move the rendering of clothes into the EntityRenderer, but that uses different matrixStack than in the FeatureRenderer and replicating that is complicated, basically the LivingEntityRenderer just sets up the matrixStack and then renders the model + features implementing the same matrixStacks makes the LivingEntityRenderer obsolete...

Another idea was to move it to the EntityModel, but unfortunately we don't have access to the entity there.

So it stays in the FeatureRenderer with minor improvements I did on the way.