Add Coin Pile item
The purpose of this item is to replace all existing coin items and use the numeric value instead.

So far it behaves like a normal item, you can merge and split the stack, but quick move and double click merge is not yet supported.

This item also has a dynamic texture that shows what coins are in the stack (this is not fully finished yet).

The tooltip shows the total value of the stack and also shows the amount of each coin if you hold the sneak button (LShift by default).

Currently the count label shows the amount of the highest coin in the stack (e.g. if the stack "has" 5 gold coins, 30 iron coins and 13 copper coins, you would see the number 5).
This may or may not change later.